pane resizing implemented
